+++ src/index.tsx
@@ -1,6 +1,3 @@
 import './modules/core/styles.css';
-export * from './modules/core';
-export * from './modules/shared';
-export * from './modules/checkout';
-export * from './modules/core';
-export * from './modules/auth';
+
+export * from './modules';