+++ src/modules/auth/components/ConnectWalletTemplate/ConnectWalletTemplate.tsx
@@ -7,7 +7,7 @@ import { Provider } from '@w3block/pixchain-react-metamask';
 import { useProfile } from '../../../shared';
 import { ReactComponent as MetamaskLogo } from '../../../shared/assets/icons/metamask.svg';
 import { Alert } from '../../../shared/components/Alert';
-import Spinner from '../../../shared/components/Spinner/Spinner';
+import { Spinner } from '../../../shared/components/Spinner';
 import TranslatableComponent from '../../../shared/components/TranslatableComponent';
 import { PixwayAPIRoutes } from '../../../shared/enums/PixwayAPIRoutes';
 import { PixwayAppRoutes } from '../../../shared/enums/PixwayAppRoutes';