+++ src/modules/tokens/interfaces/TokenCategory.ts
@@ -1,0 +1,7 @@
+export interface TokenCategory {
+  id: string;
+  createdAt: string;
+  updatedAt: string;
+  name: string;
+  image: string;
+}